snaprepoapi: Add Region field and WithRegion ctxv
